Moves copy button into its own component
