+++ custom-resources/copy-button.css
@@ -1,0 +1,3 @@
+.copy-button svg {
+    margin-top: -5px; /* Negate the empty space at the top of the button's icon */
+}
