+++ custom-resources/phoneme-picker.css
@@ -28,10 +28,6 @@
     margin-left: auto;
 }
 
-.phoneme-details .copy-button svg {
-    margin-top: -5px; /* Negate the empty space at the top of the button's icon */
-}
-
 .phoneme-detail-item {
     margin-bottom: 5px;
 }
