+++ src/index.html
@@ -14,6 +14,7 @@
     <link rel="stylesheet" href="spoken-languages.css">
     <link rel="stylesheet" href="sign-languages.css">
     <link rel="stylesheet" href="multi-select.css">
+    <link rel="stylesheet" href="copy-button.css">
   </head>
   <body>
     <div id="react-app"></div>
